tidy-up after refactor useMutationWithOptimisticUpdates
